+++ plugin-build/plugin/src/main/java/com/ncorti/ktfmt/gradle/tasks/KtfmtBaseTask.kt
@@ -54,6 +54,8 @@ abstract class KtfmtBaseTask internal constructor(private val layout: ProjectLay
     @get:Input
     abstract val includeOnly: Property<String>
 
+    @get:Input abstract val useClassloaderIsolation: Property<Boolean>
+
     @PathSensitive(PathSensitivity.RELATIVE)
     @InputFiles
     @IgnoreEmptyDirectories
@@ -87,7 +89,12 @@ abstract class KtfmtBaseTask internal constructor(private val layout: ProjectLay
     }
 
     private fun submitFilesToFormatterWorker(tmpResultDirectory: File) {
-        val queue = workerExecutor.processIsolation { it.classpath.from(ktfmtClasspath) }
+        val queue =
+            if (useClassloaderIsolation.get()) {
+                workerExecutor.classLoaderIsolation { it.classpath.from(ktfmtClasspath) }
+            } else {
+                workerExecutor.processIsolation { it.classpath.from(ktfmtClasspath) }
+            }
 
         val includedFiles = getIncludedFiles()